Moved the DOIOstiWebClient class into the osti.py module and removed the osti_web_client.py module
